chore: included feedback from review
