+++ src/util/misc.rs
@@ -10,51 +10,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use crate::args::Args;
 use super::db::delete;
 
-pub fn cleanup_orphaned_files(pastas: &Vec<Pasta>, args: &Args) {
-    log::debug!("Starting orphaned files cleanup");
-    
-    let attachments_dir = format!("{}/attachments", args.data_dir);
-    let dirs = match std::fs::read_dir(&attachments_dir) {
-        Ok(dirs) => dirs,
-        Err(e) => {
-            log::debug!("No attachments directory found or unable to read: {}", e);
-            return;
-        }
-    };
-    
-    let mut orphaned_count = 0;
-    for dir_entry in dirs.flatten() {
-        if !dir_entry.file_type().map(|ft| ft.is_dir()).unwrap_or(false) {
-            continue;
-        }
-        
-        let dir_name = dir_entry.file_name().to_string_lossy().to_string();
-        
-        // Check if any pasta references this directory
-        let is_referenced = pastas.iter().any(|pasta| {
-            pasta.file.is_some() && pasta.id_as_animals(&args.hash_ids) == dir_name
-        });
-        
-        if !is_referenced {
-            log::debug!("Found orphaned directory: {}", dir_name);
-            let full_path = dir_entry.path();
-            
-            // Remove the entire directory and its contents
-            if let Err(e) = std::fs::remove_dir_all(&full_path) {
-                log::error!("Failed to remove orphaned directory {:?}: {}", full_path, e);
-            } else {
-                log::info!("Removed orphaned directory: {:?}", full_path);
-                orphaned_count += 1;
-            }
-        }
-    }
-    
-    if orphaned_count > 0 {
-        log::info!("Orphaned files cleanup: removed {} orphaned directories", orphaned_count);
-    } else {
-        log::debug!("Orphaned files cleanup: no orphaned files found");
-    }
-}
 
 pub fn remove_expired(pastas: &mut Vec<Pasta>, args: &Args) {
     // get current time - this will be needed to check which pastas have expired
@@ -91,7 +46,6 @@ pub fn remove_expired(pastas: &mut Vec<Pasta>, args: &Args) {
                 log::debug!("  Reason: GC limit reached (last read {} days ago, limit: {})", p.last_read_days_ago(), args.gc_days);
             }
             
-            // remove from database
             delete(None, Some(p.id), args);
 
             // remove the file itself
@@ -102,7 +56,6 @@ pub fn remove_expired(pastas: &mut Vec<Pasta>, args: &Args) {
                     p.id_as_animals(&args.hash_ids),
                     file.name()
                 );
-                log::debug!("Attempting to delete file: {}", file_path);
                 if let Err(e) = fs::remove_file(&file_path) {
                     log::error!("Failed to delete file {}: {}", file_path, e);
                 } else {
@@ -115,7 +68,6 @@ pub fn remove_expired(pastas: &mut Vec<Pasta>, args: &Args) {
                     args.data_dir,
                     p.id_as_animals(&args.hash_ids)
                 );
-                log::debug!("Attempting to delete directory: {}", dir_path);
                 if let Err(e) = fs::remove_dir(&dir_path) {
                     log::error!("Failed to delete directory {}: {}", dir_path, e);
                 } else {
